% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{beta_priors}
\alias{beta_priors}
\title{Specify Prior Distributions for the slope parameters}
\usage{
beta_priors(
  k,
  beta_prior_mean = matrix(0, k, 1),
  beta_prior_var = diag(k) * 10
)
}
\arguments{
\item{k}{The total number of coefficients in the model.}

\item{beta_prior_mean}{A \eqn{k x 1} matrix of \eqn{\beta} prior means (default: vector of zeros)}

\item{beta_prior_var}{A \eqn{k x k} matrix of \eqn{\beta} prior variances (default: \eqn{10})

This function allows the user to specify priors for the slope coefficients.}
}
\description{
Specify Prior Distributions for the slope parameters
}
