% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prob.R
\name{beta_prob}
\alias{beta_prob}
\title{Return the density of the four-paramater Beta distribution for a given rho value}
\usage{
beta_prob(rho, a = 1, b = 1, rmin = 0, rmax = 1)
}
\arguments{
\item{rho}{The scalar value for \eqn{\rho}}

\item{a}{The first shape parameter of the Beta distribution}

\item{b}{The second shape parameter of the Beta distribution}

\item{rmin}{Minimum value of \eqn{\rho}}

\item{rmax}{Maximum value of \eqn{\rho}}
}
\value{
Returns the beta probability.
}
\description{
Return the density of the four-paramater Beta distribution for a given rho value
}
