% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getWknn}
\alias{getWknn}
\title{Constructs a nearest neighbour spatial weight matrix using lat/long coordinates and geodesic distance.}
\usage{
getWknn(xy, k)
}
\arguments{
\item{xy}{An \eqn{n x 2} matrix of x and y coordinates.}

\item{k}{Number of nearest neighbours}
}
\value{
A sparse \eqn{n x n} spatial weight matrix
}
\description{
Constructs a nearest neighbour spatial weight matrix using lat/long coordinates and geodesic distance.
}
