% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{init_priors}
\alias{init_priors}
\title{Specify Prior Distributions for Estimate W Models}
\usage{
init_priors(
  X,
  Z = NULL,
  tt = 1,
  W_prior = matrix(0.5, nrow(X)/tt, nrow(X)/tt),
  beta_prior_mean = matrix(0, ifelse(is.null(X), 0, ncol(X)) * 2 + ifelse(is.null(Z),
    0, ncol(Z)), 1),
  beta_prior_var = diag(ifelse(is.null(X), 0, ncol(X)) * 2 + ifelse(is.null(Z), 0,
    ncol(Z))) * 10,
  sigma_a = 0.1,
  sigma_b = 0.1,
  rho_pr = 1.1,
  griddy_n = 105,
  rmin = 0,
  rmax = 1,
  bbinom_a = 1,
  bbinom_b = 1,
  bb_pr = TRUE,
  min_k = 1,
  max_k = min(10, floor(nrow(X)/tt/2)),
  rjct_pr = FALSE
)
}
\arguments{
\item{X}{An \eqn{n x q_x} matrix of independent variables that will spatially lagged (can be NULL)}

\item{Z}{An \eqn{n x q_z} matrix of independent variables that will not be spatially lagged}

\item{tt}{Number of time observations}

\item{W_prior}{An \eqn{n x n} matrix of priors for \eqn{W}}

\item{beta_prior_mean}{A \eqn{k x 1} matrix of \eqn{\beta} prior means}

\item{beta_prior_var}{A \eqn{k x k} matrix of \eqn{\beta} prior variances}

\item{sigma_a}{Sigma rate prior parameter (scalar)}

\item{sigma_b}{Sigma shape prior parameter (scalar)}

\item{rho_pr}{Prior for the four-part beta ditribution}

\item{griddy_n}{Number of griddy gibbs steps}

\item{rmin}{Minimum range of \eqn{\rho}}

\item{rmax}{Maximum range of \eqn{\rho}}

\item{bbinom_a}{Parameter a of sparsity prior}

\item{bbinom_b}{Parameter a of sparsity prior}

\item{bb_pr}{Should sparsity priors be used? (default: TRUE)}

\item{min_k}{Minimum number of a priori neighbours}

\item{max_k}{Maximum number of a priori neighbours}

\item{rjct_pr}{Should rejection priors be used? (default: TRUE)

This function gives access to a larger set of prior distributions in case the default choice is unsatisfactory.}
}
\description{
Specify Prior Distributions for Estimate W Models
}
