% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_util.R
\name{init_sampler_W}
\alias{init_sampler_W}
\title{Initialize \code{wdraws} object for tracking MCMC draws of the spatial weight matrix}
\usage{
init_sampler_W(W_prior, rho = NULL)
}
\arguments{
\item{W_prior}{Priors for the spatial weight matrix, generated by \code{\link{W_priors}}}

\item{rho}{The value of the spatial autoregressive parameter. If \code{NULL} (default) then no log determinants or inverses of \eqn{I-\rho W} are generated.}
}
\value{
An object of class \code{wdraws}
}
\description{
Initialize \code{wdraws} object for tracking MCMC draws of the spatial weight matrix
}
