% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lndetPaceBarry.R
\name{lndetPaceBarry}
\alias{lndetPaceBarry}
\title{Pace and Barry's log determinant approximation}
\usage{
lndetPaceBarry(W, length.out = 200, rmin = -1, rmax = 1)
}
\arguments{
\item{W}{An \eqn{n x n} spatial weights matrix}

\item{length.out}{Length of grid (defaults to 200)}

\item{rmin}{Minimum range of \eqn{\rho} (defaults to -1)}

\item{rmax}{Minimum \eqn{\rho} value (defaults to 1)}
}
\value{
A \eqn{length.out x 2} matrix; the first column contains the approximated log-determinants
the second column the rho values
}
\description{
Pace and Barry's log determinant approximation
}
