% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngamma.R
\name{normalgamma}
\alias{normalgamma}
\title{A sampler for a linear model with independent Normal-Gamma priors}
\usage{
normalgamma(
  Y,
  tt,
  X = matrix(1, nrow(Y), 1),
  niter = 1000,
  nretain = 250,
  beta_prior = beta_priors(k = ncol(X)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{numeric \eqn{tt * n} by \eqn{1} matrix containing the dependent variables.}

\item{tt}{single number greater or equal to 1. Denotes the number of time observations.}

\item{X}{numeric \eqn{tt * n} by \eqn{k_1} design matrix of independent variables.}

\item{niter}{single number greater or equal to 1, indicating the total number of draws.
Will be automatically coerced to integer. The default value is 1000.}

\item{nretain}{single number greater or equal to 0, indicating the number of draws
kept after the burn-in. Will be automatically coerced to integer. The default value is 500.}

\item{beta_prior}{list containing priors for the slope coefficients \eqn{\beta},
generated by the smart constructor \code{\link{beta_priors}}.}

\item{sigma_prior}{list containing priors for the error variance \eqn{\sigma^2},
generated by the smart constructor \code{\link{sigma_priors}}}
}
\description{
The sampler uses independent Normal-Gamma priors to estimate a linear model. The function is
used for an illustration on using the \code{\link{beta_sampler}} and \code{\link{sigma_sampler}}
classes.
}
\examples{
n = 20; tt = 10; k = 3
X = matrix(stats::rnorm(n*tt*k),n*tt,k)
Y = X \%*\% c(1,0,-1) + stats::rnorm(n*tt,0,.5)
res = normalgamma(Y,tt,X)
}
