% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts1growth}
\alias{nuts1growth}
\title{Regional growth data}
\format{
A \code{data.frame} object.
}
\usage{
nuts1growth
}
\description{
Regional growth data set contains information on annual growth rates of GVA per worker
(labor productivity) for 90 European NUTS-1 regions for the period 1999-2019.
}
\details{
Data is provided for NUTS-1 regions of the following countries: ...

The dataset includes growth of GDP per worker
}
\references{
Piribauer, P., Glocker, C., & Krisztin, T. (2023). Beyond distance: The spatial relationships of European regional economic growth. Journal of Economic Dynamics and Control, 155, 104735.
}
\keyword{growth}
\keyword{regional}
