% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.estimateW}
\alias{plot.estimateW}
\title{Graphical Summary of the estimated W matrix}
\usage{
\method{plot}{estimateW}(
  x,
  cols = c("white", "lightgrey", "black"),
  breaks = c(0, 0.5, 0.75, 1),
  main = "Posterior incl. prob. of W",
  ...
)
}
\arguments{
\item{x}{\code{estimateW} object.}

\item{cols}{Main colors to use for the plot}

\item{breaks}{Breaks for the colours}

\item{main}{Legend title}

\item{\dots}{further arguments are passed on to be invoked}
}
\description{
Graphical Summary of the estimated W matrix
}
