% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.sim_dgp}
\alias{plot.sim_dgp}
\title{Graphical summary of a generated spatial weight matrix}
\usage{
\method{plot}{sim_dgp}(x, ...)
}
\arguments{
\item{x}{\code{sim_dgp} object}

\item{\dots}{further arguments are passed on to the invoked}
}
\description{
Graphical summary of a generated spatial weight matrix
}
