% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_samplers.R
\docType{class}
\name{rho_samplerMH}
\alias{rho_samplerMH}
\title{An R6 class for sampling \eqn{\rho} using a Metropolis-Hastings step}
\format{
An \code{\link{R6Class}} generator object
}
\description{
An R6 class for sampling \eqn{\rho} using a Metropolis-Hastings step

An R6 class for sampling \eqn{\rho} using a Metropolis-Hastings step
}
\details{
This class samples the spatial autoregressive coefficient. Use the \link{rho_priors} class for setup.

The sampling relies on a Metropolis-Hastings step, as outlined in LeSage and Pace (2009).
}
\references{
LeSage, J. and Pace, R.K., (2009) Introduction to spatial econometrics. Chapman and Hall/CRC.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rho_prior}}{The current \code{\link{rho_priors}}}

\item{\code{curr_rho}}{The current value of \eqn{\rho}}

\item{\code{curr_W}}{The current value of the spatial weight matrix \eqn{W}; an \eqn{n} by \eqn{n} matrix.}

\item{\code{curr_A}}{The current spatial projection matrix \eqn{I - \rho W}.}

\item{\code{curr_AI}}{The inverse of \code{curr_A}}

\item{\code{curr_logdet}}{The current log-determinant of \code{curr_A}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rho_samplerMH-new}{\code{rho_samplerMH$new()}}
\item \href{#method-rho_samplerMH-stopMHtune}{\code{rho_samplerMH$stopMHtune()}}
\item \href{#method-rho_samplerMH-setW}{\code{rho_samplerMH$setW()}}
\item \href{#method-rho_samplerMH-sample}{\code{rho_samplerMH$sample()}}
\item \href{#method-rho_samplerMH-clone}{\code{rho_samplerMH$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rho_samplerMH-new"></a>}}
\if{latex}{\out{\hypertarget{method-rho_samplerMH-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rho_samplerMH$new(rho_prior, W = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rho_prior}}{The list returned by \code{\link{rho_priors}}}

\item{\code{W}}{An optional starting value for the spatial weight matrix \eqn{W}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rho_samplerMH-stopMHtune"></a>}}
\if{latex}{\out{\hypertarget{method-rho_samplerMH-stopMHtune}{}}}
\subsection{Method \code{stopMHtune()}}{
Function to stop the tuning of the Metropolis-Hastings step. The tuning of the
Metropolis-Hastings step is usually carried out until half of the burn-in phase.
Call this function to turn it off.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rho_samplerMH$stopMHtune()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rho_samplerMH-setW"></a>}}
\if{latex}{\out{\hypertarget{method-rho_samplerMH-setW}{}}}
\subsection{Method \code{setW()}}{
If the spatial weight matrix is updated during the sampling procedure the log determinant, the
spatial projection matrix \eqn{I - \rho W} and it's inverse must be updated. This function should be
used for a consistent update. At least the new spatial weight matrix \code{newW} should be supplied.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rho_samplerMH$setW(newW, newLogdet = NULL, newA = NULL, newAI = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newW}}{The updated spatial weight matrix \eqn{W}}

\item{\code{newLogdet}}{An optional value for the log determinant corresponding to \çode{newW} and \code{curr_rho}}

\item{\code{newA}}{An optional value for the spatial projection matrix using \çode{newW} and \code{curr_rho}}

\item{\code{newAI}}{An optional value for the matrix inverse of \code{newA}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rho_samplerMH-sample"></a>}}
\if{latex}{\out{\hypertarget{method-rho_samplerMH-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rho_samplerMH$sample(Y, mu, sigma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{The \eqn{n} by \eqn{tt} vector of responses}

\item{\code{mu}}{The \eqn{n} by \eqn{tt} vector of means (usually \eqn{X\beta})}

\item{\code{sigma}}{The variance parameter \eqn{\sigma^2}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rho_samplerMH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rho_samplerMH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rho_samplerMH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
