% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleW_fast}
\alias{sampleW_fast}
\title{A fast sampling step implemented in C++ for updating the spatial weight matrix.}
\usage{
sampleW_fast(
  Y,
  curr_sigma,
  mu,
  lag_mu,
  W_prior,
  curr_W,
  curr_w,
  curr_A,
  curr_AI,
  curr_logdet,
  curr_rho,
  nr_neighbors_prior,
  symmetric,
  spatial_error,
  row_standardized
)
}
\arguments{
\item{Y}{The \eqn{n} by \eqn{tt} matrix of responses}

\item{curr_sigma}{The variance parameter \eqn{\sigma^2}}

\item{mu}{The \eqn{n} by \eqn{tt} matrix of means.}

\item{lag_mu}{\eqn{n} by \eqn{tt} matrix of means that will be spatially lagged with
the estimated \eqn{W}. Defaults to a matrix with zero elements.}

\item{W_prior}{The current \code{\link{W_priors}}}

\item{curr_W}{binary \eqn{n} by \eqn{n} spatial connectivity matrix \eqn{\Omega}}

\item{curr_w}{Row-standardized \eqn{n} by \eqn{n} spatial weight matrix \eqn{W}}

\item{curr_A}{The current spatial projection matrix \eqn{I - \rho W}.}

\item{curr_AI}{The inverse of \code{curr_A}}

\item{curr_logdet}{The current log-determinant of \code{curr_A}}

\item{curr_rho}{single number between -1 and 1 or NULL, depending on whether the sampler updates
the spatial autoregressive parameter \eqn{\rho}. Set while invoking \code{initialize}
or using the function \code{set_rho}.}

\item{nr_neighbors_prior}{An \eqn{n} dimensional vector of prior weights on the number of neighbors
(i.e. the row sums of the adjacency matrix \eqn{\Omega}), where the first element denotes the prior probability
of zero neighbors and the last those of \eqn{n-1}. A prior using only fixed inclusion probabilities
for the entries in \eqn{\Omega} would be an \eqn{n} dimensional vector of \eqn{1/n}. Defaults to
a \code{\link{bbinompdf}} prior, with prior parameters \eqn{a = 1}, \eqn{b = 1}.}

\item{symmetric}{Binary value. Should the estimated adjacency matrix \eqn{\Omega} be symmetric (default: FALSE)?
if TRUE: \eqn{\Omega} is forced symmetric; if FALSE: \eqn{\Omega} not necessarily symmetric.}

\item{spatial_error}{Should a spatial error model be constructed? Defaults to \code{FALSE}.}

\item{row_standardized}{Binary value. Should the estimated \eqn{W} matrix be row-standardized (default: TRUE)?
if TRUE: row-stochastic \eqn{W}; if FALSE: \eqn{W} not row-standardized.}
}
\description{
This function is intended to be called from the R6 class \code{\link{W_sampler}}.
}
