% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR_W.R
\name{SAR_W}
\alias{SAR_W}
\title{A sampler for estimating the W matrix in a SAR type model \eqn{Y = \rho WY + X \beta + \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}}
\usage{
SAR_W(
  Y,
  X,
  W_prior,
  tt,
  niter = 1000,
  nretain = 250,
  beta_prior_mean = matrix(0, ncol(X), 1),
  beta_prior_var = diag(ncol(X)) * 10,
  sigma_a = 0.1,
  sigma_b = 0.1,
  rho_pr = 1.1,
  griddy_n = 105,
  rmin = 0,
  rmax = 1,
  bbinom_a = 1,
  bbinom_b = 1,
  bb_pr = TRUE,
  min_k = 1,
  max_k = min(10, floor(nrow(W_prior)/2)),
  rjct_pr = FALSE,
  VERBOSE = FALSE,
  SYMMETRIC = TRUE,
  GRIDDY_GIBBS = TRUE,
  ROW_STANDARDIZED = TRUE
)
}
\arguments{
\item{Y}{An \eqn{n x 1} matrix of dependent variables}

\item{X}{An \eqn{n x k} matrix of independent variables}

\item{W_prior}{An \eqn{n x x} matrix of priors for \eqn{W}}

\item{tt}{Number of time observations}

\item{niter}{Total number of iterations}

\item{nretain}{Number of iterations to retain (must be smaller than niter)}

\item{beta_prior_mean}{A \eqn{k x 1} matrix of \eqn{\beta} prior means}

\item{beta_prior_var}{A \eqn{k x k} matrix of \eqn{\beta} prior variances}

\item{sigma_a}{Sigma rate prior parameter (scalar)}

\item{sigma_b}{Sigma shape prior parameter (scalar)}

\item{rho_pr}{Prior for the four-part beta ditribution}

\item{griddy_n}{Number of griddy gibbs steps}

\item{rmin}{Minimum range of \eqn{\rho}}

\item{rmax}{Maximum range of \eqn{\rho}}

\item{bbinom_a}{Parameter a of sparsity prior}

\item{bbinom_b}{Parameter a of sparsity prior}

\item{bb_pr}{Should sparsity priors be used? (default: TRUE)}

\item{min_k}{Minimum number of a priori neighbours}

\item{max_k}{Maximum number of a priori neighbours}

\item{rjct_pr}{Should rejection priors be used? (default: TRUE)}

\item{VERBOSE}{Should detailed diagnostic information be provided (default: FALSE)}

\item{SYMMETRIC}{Should the estimated \eqn{W} matrix be symmetric (default: TRUE)}

\item{GRIDDY_GIBBS}{Should griddy-Gibbs be used for \eqn{\rho} estimation?
Does not work if \code{ROW_STANDARDIZED = FALSE}. Main advantage is that less draws are required for \eqn{\rho}}

\item{ROW_STANDARDIZED}{Should the estimated \eqn{W} matrix be row-standardized (default: TRUE)}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and direct, indirect, and total effects.
}
\description{
A sampler for estimating the W matrix in a SAR type model \eqn{Y = \rho WY + X \beta + \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}
}
\examples{
n = 20; tt = 10
dgp_dat = sarwsim(n =n, tt = 10, k=3, rho = .5, beta = c(1,-1,1), sigma2 = .5)
res = SAR_W(Y = dgp_dat$Y,X = dgp_dat$X,tt = tt,W_prior = matrix(.5,n,n),niter = 100,nretain = 50)
}
