% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmw.R
\name{sarw}
\alias{sarw}
\title{A sampler for estimating the W matrix in a SAR type model}
\usage{
sarw(
  Y,
  tt,
  Z,
  niter = 1000,
  nretain = 250,
  W_prior = W_priors(n = nrow(Y)/tt),
  rho_prior = rho_priors(),
  beta_prior = beta_priors(k = ncol(Z)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{numeric \eqn{tt * n} by \eqn{1} matrix containing the dependent variables.}

\item{tt}{single number greater or equal to 1. Denotes the number of time observations.}

\item{Z}{numeric \eqn{tt * n} by \eqn{k_3} design matrix of independent variables (which are not
spatially lagged by \eqn{W}). The default value
is a \eqn{tt * n} by \eqn{1} vector of ones (an intercept for the model).}

\item{niter}{single number greater or equal to 1, indicating the total number of draws.
Will be automatically coerced to integer. The default value is 1000.}

\item{nretain}{single number greater or equal to 0, indicating the number of draws
kept after the burn-in. Will be automatically coerced to integer. The default value is 500.}

\item{W_prior}{list containing prior settings for estimating the spatial weight matrix \eqn{W},
generated by the smart constructor \code{\link{W_priors}}.}

\item{rho_prior}{list of prior settings for estimating \eqn{\rho},
generated by the smart constructor \code{\link{rho_priors}}}

\item{beta_prior}{list containing priors for the slope coefficients \eqn{\beta},
generated by the smart constructor \code{\link{beta_priors}}.}

\item{sigma_prior}{list containing priors for the error variance \eqn{\sigma^2},
generated by the smart constructor \code{\link{sigma_priors}}}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and direct, indirect, and total effects.
}
\description{
### TODO
The model takes the form \eqn{Y = \rho f(\Omega)Y + Z \beta +  \epsilon}, with \eqn{\epsilon \sim N(0,I \sigma^2)}
}
\details{
This is a wrapper function calling \code{\link{sdmw}} with no spatially lagged dependent variables.
}
\examples{
n = 20; tt = 10
dgp_dat = sim_dgp(n =n, tt = tt, rho = .5, beta3 = c(.5,1), sigma2 = .5)
res = sarw(Y = dgp_dat$Y,tt = tt,Z = dgp_dat$Z,niter = 20,nretain = 10)
}
