% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sarwsim}
\alias{sarwsim}
\title{Simulating a SAR data generating process}
\usage{
sarwsim(n, tt, k, rho, beta, sigma2 = 0.5, n_neighbor = 7, SYMMETRIC = FALSE)
}
\arguments{
\item{n}{Number of spatial observations}

\item{tt}{Number of time observations}

\item{k}{Number of explanatory variables}

\item{rho}{The true rho parameter}

\item{beta}{The \eqn{k x 1} vector of true betas}

\item{sigma2}{The true \eqn{\sigma^2} parameter}

\item{n_neighbor}{Number of neighbours for the \eqn{W} matrix}

\item{SYMMETRIC}{Should the generated \eqn{W} matrix be symmetric? (default: FALSE)}
}
\value{
A list with the generated \eqn{X}, \eqn{Y} and \eqn{W} and a list of parameters.
}
\description{
Simulating a SAR data generating process
}
\examples{
dgp_dat = sarwsim(n =20, tt = 10, k=3, rho = .5, beta = c(1,-1,1), sigma2 = .5)

}
