% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmw.R
\name{sdmw}
\alias{sdmw}
\title{A sampler for estimating the W matrix in a SAR type model}
\usage{
sdmw(
  Y,
  X = matrix(0, nrow(Y), 0),
  Z = matrix(0, nrow(Y), 0),
  tt,
  niter = 1000,
  nretain = 250,
  priors = init_priors(X, Z, tt),
  VERBOSE = FALSE,
  SYMMETRIC = TRUE,
  GRIDDY_GIBBS = TRUE,
  ROW_STANDARDIZED = TRUE
)
}
\arguments{
\item{Y}{An \eqn{tt n x 1} matrix of dependent variables}

\item{X}{An \eqn{tt n x q_x} matrix of independent variables}

\item{Z}{An \eqn{tt n x q_k} matrix of independent variables}

\item{tt}{Number of time observations}

\item{niter}{Total number of iterations}

\item{nretain}{Number of iterations to retain (must be smaller than niter)}

\item{priors}{List of priors, generated by \code{\link{init_priors}}}

\item{VERBOSE}{Should detailed diagnostic information be provided (default: FALSE)}

\item{SYMMETRIC}{Should the estimated \eqn{W} matrix be symmetric (default: TRUE)}

\item{GRIDDY_GIBBS}{Should griddy-Gibbs be used for \eqn{\rho} estimation?
Does not work if \code{ROW_STANDARDIZED = FALSE}. Main advantage is that less draws are required for \eqn{\rho}}

\item{ROW_STANDARDIZED}{Should the estimated \eqn{W} matrix be row-standardized (default: TRUE)}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and direct, indirect, and total effects.
}
\description{
The model takes thes form \eqn{Y = \rho WY + X \beta_1 + WX \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}
}
\examples{
n = 20; tt = 10
dgp_dat = sarwsim(n =n, tt = 10, k=3, rho = .5, beta = c(1,-1,1), sigma2 = .5)
res = sdmw(Y = dgp_dat$Y,X = matrix(0,n*tt,0),Z = dgp_dat$X,tt = tt,niter = 20,nretain = 10)
}
