% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_samplers.R
\docType{class}
\name{sigma_sampler}
\alias{sigma_sampler}
\title{An R6 class for sampling for sampling \eqn{\sigma^2}}
\format{
An \code{\link{R6Class}} generator object
}
\description{
This class samples nuisance parameter which an inverse Gamma prior from the conditional posterior.
Use the \link{sigma_priors} class for setup.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sigma_prior}}{The current \code{\link{sigma_priors}}}

\item{\code{curr_sigma}}{The current value of \eqn{\beta}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sigma_sampler-new}{\code{sigma_sampler$new()}}
\item \href{#method-sigma_sampler-sample}{\code{sigma_sampler$sample()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sigma_sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-sigma_sampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sigma_sampler$new(sigma_prior)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sigma_prior}}{The list returned by \code{\link{sigma_priors}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sigma_sampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sigma_sampler-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sigma_sampler$sample(Y, mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{The \eqn{n} by \eqn{1} matrix of responses}

\item{\code{mu}}{The \eqn{n} by \eqn{1} matrix of means}
}
\if{html}{\out{</div>}}
}
}
}
