% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_dgp}
\alias{sim_dgp}
\title{Simulating from an SDM data generating process}
\usage{
sim_dgp(
  n,
  tt,
  rho,
  beta1 = c(),
  beta2 = c(),
  beta3 = c(),
  sigma2 = 0.5,
  n_neighbor = 7,
  do_symmetric = FALSE
)
}
\arguments{
\item{n}{Number of spatial observations}

\item{tt}{Number of time observations}

\item{rho}{The true rho parameter}

\item{beta1}{Numeric vector of length \eqn{k_1} vector. Provides the values for \eqn{\beta_1} Defaults
to \code{c()}.}

\item{beta2}{Numeric vector of length \eqn{k_1} vector. Provides the values for \eqn{\beta_2} Defaults
to \code{c()}.}

\item{beta3}{Numeric vector of length \eqn{k_3} vector. Provides the values for \eqn{\beta_3} Defaults
to \code{c()}.}

\item{sigma2}{The true \eqn{\sigma^2} parameter}

\item{n_neighbor}{Number of neighbors for the \eqn{W} matrix}

\item{do_symmetric}{Should the generated \eqn{W} matrix be symmetric? (default: FALSE)}
}
\value{
A list with the generated \eqn{X}, \eqn{Y} and \eqn{W} and a list of parameters.
}
\description{
The model takes the form \eqn{Y = \rho W Y + X \beta_1 + W X \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,I\sigma^2)}
}
\examples{
# SDM data generating process
dgp_dat = sim_dgp(n =20, tt = 10, rho = .5, beta1 = c(1,-1),
                  beta2 = c(0,.5),beta3 = c(.2),sigma2 = .5)
}
