% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_sdmw}
\alias{sim_sdmw}
\title{Simulating an SDM data generating process}
\usage{
sim_sdmw(
  n,
  tt,
  rho,
  beta1 = c(),
  beta2 = c(),
  beta3 = c(),
  sigma2 = 0.5,
  n_neighbor = 7,
  do_symmetric = FALSE
)
}
\arguments{
\item{n}{Number of spatial observations}

\item{tt}{Number of time observations}

\item{rho}{The true rho parameter}

\item{beta1}{The \eqn{k_1 x 1} vector of true \eqn{\beta_1}}

\item{beta2}{The \eqn{k_1 x 1} vector of true \eqn{\beta_2} (with \eqn{k_1 = k_2})}

\item{beta3}{The \eqn{k_3 x 1} vector of true \eqn{\beta_3}}

\item{sigma2}{The true \eqn{\sigma^2} parameter}

\item{n_neighbor}{Number of neighbours for the \eqn{W} matrix}

\item{do_symmetric}{Should the generated \eqn{W} matrix be symmetric? (default: FALSE)}
}
\value{
A list with the generated \eqn{X}, \eqn{Y} and \eqn{W} and a list of parameters.
}
\description{
The model takes the form \eqn{Y = \rho W Y + X \beta_1 + W X \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}
}
\examples{
# SDM data generating process
dgp_dat = sim_sdmw(n =20, tt = 10, rho = .5, beta1 = c(1,-1), beta2 = c(0,.5),beta3 = c(.2),sigma2 = .5)
}
