% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ldetAI.R
\name{update_ldetAI}
\alias{update_ldetAI}
\title{Efficient update of the log-determinant and the matrix inverse using the
Matrix Determinant Lemma and the Sherran-Morrison formula}
\usage{
update_ldetAI(ch_ind, diff, AI, logdet)
}
\arguments{
\item{ch_ind}{The index of rows to update}

\item{diff}{A \eqn{ch_ind x n} matrix. This value will be added to the corresponding rows of \eqn{A}.}

\item{AI}{A \eqn{n x n} inverse of the matrix \eqn{A}, where \eqn{A = I - \rho W}.}

\item{logdet}{The log-determinant of \eqn{A}.}
}
\value{
A list containing \eqn{A^-1} and the log determinant of \eqn{A}
}
\description{
Efficient update of the log-determinant and the matrix inverse using the
Matrix Determinant Lemma and the Sherran-Morrison formula
}
