% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{W_priors}
\alias{W_priors}
\title{Specify Prior Distributions for the spatial weight matrix}
\usage{
W_priors(
  n,
  W_prior = matrix(0.5, n, n),
  SYMMETRIC = TRUE,
  ROW_STANDARDIZED = TRUE,
  bbinom_a = 1,
  bbinom_b = 1,
  bb_pr = TRUE,
  min_k = 1,
  max_k = min(10, floor(n/2)),
  rjct_pr = FALSE
)
}
\arguments{
\item{n}{The number of observations}

\item{W_prior}{An \eqn{n x n} matrix of priors for \eqn{W}}

\item{SYMMETRIC}{Should the estimated \eqn{W} matrix be symmetric (default: TRUE)}

\item{ROW_STANDARDIZED}{Should the estimated \eqn{W} matrix be row-standardized (default: TRUE)}

\item{bbinom_a}{Parameter a of sparsity prior}

\item{bbinom_b}{Parameter a of sparsity prior}

\item{bb_pr}{Should sparsity priors be used? (default: TRUE)}

\item{min_k}{Minimum number of a priori neighbours}

\item{max_k}{Maximum number of a priori neighbours}

\item{rjct_pr}{Should rejection priors be used? (default: TRUE)

This function gives access to a larger set of prior distributions in case the default choice is unsatisfactory.}
}
\description{
Specify Prior Distributions for the spatial weight matrix
}
