% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbinompdf.R
\name{bbinompdf}
\alias{bbinompdf}
\title{The beta binomial probability density function for sparsity priors}
\usage{
bbinompdf(x, nsize, a, b, min_k = 0, max_k = nsize)
}
\arguments{
\item{x}{Number of neighbors (scalar)}

\item{nsize}{Maximal number of elements}

\item{a}{Scalar prior parameter a}

\item{b}{Scalar prior parameter b}

\item{min_k}{Minimum number of elements (defaults to 0)}

\item{max_k}{Maximum number of elements (defaults to nsize)}
}
\value{
Resulting prior density of evaluated at \code{x}.
}
\description{
The beta binomial distribution can be used in \code{\link{W_priors}} as a prior on the
number of expected neighbors. Assuming a \emph{fixed} prior for \eqn{\Omega}  implies
that the number of neighbors follows a Binomial distribution with a prior
expected number of neighbors of \eqn{(N-1)\underline{p}}. However, this prior structure
can promote a relatively large number of neighbors. To put more prior weight on
parsimonious neighborhood structures and  promote sparsity in \eqn{\Omega},
the beta binomial prior accounts for the number of linkages in each row of \eqn{\Omega}.
}
\details{
### TODO: REWRITE
This flexible prior structure on the number of neighbors is a truncated beta-binomial distribution
with two prior hyperparameters \code{a} and \code{b}. The resulting prior on \eqn{\omega_{ij}} can be written as:

 \deqn{
 p(\omega_{ij} = 1 | x)\propto \Gamma\left(a+ x \right)\Gamma\left(b+(N-1)-x\right),
 }

 where \eqn{\Gamma(\cdot )} is the Gamma function, and \eqn{a} and
 \eqn{b} are prior hyperparameters. In the case of \eqn{a = b = 1}, the prior takes the
 form of a discrete uniform distribution over the number of neighbors. By fixing \eqn{a = 1}
 the prior can be anchored around the expected number of neighbors \eqn{m} through
 \eqn{b=[(N-1)-m]/m} (see Ley and Steel, 2009).

 The prior is truncated when \code{x < min_k} or when \code{x > max_k}.
}
\references{
Ley, E., & Steel, M. F. (2009). On the effect of prior assumptions in Bayesian
  model averaging with applications to growth regression. \emph{Journal of Applied Econometrics},
  \bold{24(4)}. \doi{10.1002/jae.1057}.
}
