% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prob.R
\name{beta_prob}
\alias{beta_prob}
\title{Return the density of the four-paramater Beta distribution for a given rho value}
\usage{
beta_prob(rho, a)
}
\arguments{
\item{rho}{The scalar value for \eqn{\rho}}

\item{a}{The distribution parameter; see LeSage and Pace (2019)}
}
\value{
Returns the beta probability.
}
\description{
Return the density of the four-paramater Beta distribution for a given rho value
}
