% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.estimateW}
\alias{plot.estimateW}
\title{Graphical Summary of the estimated W matrix}
\usage{
\method{plot}{estimateW}(x, ...)
}
\arguments{
\item{x}{\code{estimateW} object.}

\item{\dots}{further arguments are passed on to the invoked}
}
\description{
Graphical Summary of the estimated W matrix
}
