% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.estimateW}
\alias{plot.estimateW}
\title{Graphical summary of the estimated adjacency matrix \eqn{\Omega}}
\usage{
\method{plot}{estimateW}(
  x,
  cols = c("white", "lightgrey", "black"),
  breaks = c(0, 0.5, 0.75, 1),
  ...
)
}
\arguments{
\item{x}{\code{estimateW} object.}

\item{cols}{Main colors to use for the plot}

\item{breaks}{Breaks for the colors}

\item{\dots}{further arguments are passed on to be invoked}
}
\description{
Graphical plot of the posterior probabilities of the estimated adjacency matrix \eqn{\Omega}.
}
