% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_util.R
\name{sample_W}
\alias{sample_W}
\title{Sample \eqn{W} based on the current state of the the sampler}
\usage{
sample_W(
  tY,
  curr.txb1 = matrix(0, nrow(tY), ncol(tY)),
  curr.txb2 = matrix(0, nrow(tY), ncol(tY)),
  curr.sigma,
  W_prior,
  wdraws,
  curr.rho = NULL
)
}
\arguments{
\item{tY}{An \eqn{n x tt} matrix of dependent variables}

\item{curr.txb1}{An \eqn{n x tt} matrix of means based on \eqn{X \beta_1}}

\item{curr.txb2}{An \eqn{n x tt} matrix of means that will be spatially lagged based on \eqn{X \beta_2}}

\item{curr.sigma}{The variance parameter}

\item{W_prior}{Priors for the spatial weight matrix, generated by \code{\link{W_priors}}}

\item{wdraws}{An object of class \code{wdraws}}

\item{curr.rho}{The value of the spatial autoregressive parameter. If \code{NULL} (default) then no log determinants or inverses of \eqn{I-\rho W} are generated.}
}
\value{
An object of class \code{wdraws}
}
\description{
Sample \eqn{W} based on the current state of the the sampler
}
