% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngamma.R
\name{sar}
\alias{sar}
\title{A sampler for the Spatial Autoregressive Model with exogeneous spatial weight matrix.}
\usage{
sar(
  Y,
  tt,
  W,
  Z = matrix(1, nrow(Y), 1),
  niter = 1000,
  nretain = 250,
  rho_prior = rho_priors(),
  beta_prior = beta_priors(k = ncol(Z)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{numeric \eqn{tt * n} by \eqn{1} matrix containing the dependent variables.}

\item{tt}{single number greater or equal to 1. Denotes the number of time observations.}

\item{W}{numeric, non-negative \eqn{n} by \eqn{n} exogenous spatial weight matrix. Must have
zeros on the main diagonal.}

\item{Z}{numeric \eqn{tt * n} by \eqn{k_3} design matrix of independent variables. The default value
is a \eqn{tt * n} by \eqn{1} vector of ones (an intercept for the model).}

\item{niter}{single number greater or equal to 1, indicating the total number of draws.
Will be automatically coerced to integer. The default value is 1000.}

\item{nretain}{single number greater or equal to 0, indicating the number of draws
kept after the burn-in. Will be automatically coerced to integer. The default value is 500.}

\item{rho_prior}{list of prior settings for estimating \eqn{\rho},
generated by the smart constructor \code{\link{rho_priors}}}

\item{beta_prior}{list containing priors for the slope coefficients \eqn{\beta},
generated by the smart constructor \code{\link{beta_priors}}.}

\item{sigma_prior}{list containing priors for the error variance \eqn{\sigma^2},
generated by the smart constructor \code{\link{sigma_priors}}}
}
\description{
The sampler uses independent an Normal-Gamma prior for the slope and variance parameters,
as well as the four-parameter prior for the spatial autoregressive parameter. The function is
meant intended as an illustration on using the \code{\link{beta_sampler}}, \code{\link{sigma_sampler}},
and \code{\link{rho_sampler}} classes.
}
\details{
This is a wrapper function calling \code{\link{sdm}} with no spatially lagged dependent variables.
}
\examples{
n = 20; tt = 10
dgp_dat = sim_dgp(n =n, tt = tt, rho = .5, beta3 = c(1,.5), sigma2 = .5)
res = sar(Y = dgp_dat$Y,tt = tt, W = dgp_dat$W,
          Z = dgp_dat$Z,niter = 200,nretain = 100)
}
