% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmw.R
\name{sarw}
\alias{sarw}
\title{A sampler for estimating the W matrix in a SAR type model}
\usage{
sarw(
  Y,
  tt,
  X,
  niter = 1000,
  nretain = 250,
  W_prior = W_priors(n = nrow(Y)/tt),
  rho_prior = rho_priors(),
  beta_prior = beta_priors(k = ncol(X)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{An \eqn{tt n x 1} matrix of dependent variables}

\item{tt}{Number of time observations}

\item{X}{An \eqn{tt n x q_x} matrix of independent variables}

\item{niter}{Total number of iterations}

\item{nretain}{Number of iterations to retain (must be smaller than niter)}

\item{W_prior}{List of priors for estimating \eqn{W}, generated by \code{\link{W_priors}}}

\item{rho_prior}{List of priors for estimating \eqn{\rho}, generated by \code{\link{rho_priors}}}

\item{beta_prior}{List of priors for the slope coefficients \eqn{\beta}, generated by \code{\link{beta_priors}}}

\item{sigma_prior}{List of priors for the error variance \eqn{\sigma^2}, generated by \code{\link{sigma_priors}}}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and direct, indirect, and total effects.
}
\description{
The model takes the form \eqn{Y = \rho f(\Omega)Y + X \beta_1 +  \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}
}
\examples{
n = 20; tt = 10
dgp_dat = sim_sarw(n =n, tt = tt, rho = .5, beta1 = c(.5,1), sigma2 = .5)
res = sarw(Y = dgp_dat$Y,tt = tt,X = dgp_dat$X,niter = 20,nretain = 10)
}
