% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmw.R
\name{sdmw}
\alias{sdmw}
\title{A sampler for estimating the W matrix in a SDM type model}
\usage{
sdmw(
  Y,
  tt,
  X = matrix(0, nrow(Y), 0),
  Z = matrix(1, nrow(Y), 1),
  niter = 1000,
  nretain = 250,
  W_prior = W_priors(n = nrow(Y)/tt),
  rho_prior = rho_priors(),
  beta_prior = beta_priors(k = ncol(X) * 2 + ncol(Z)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{An \eqn{tt n x 1} matrix of dependent variables}

\item{tt}{Number of time observations}

\item{X}{An \eqn{tt n x q_x} matrix of independent variables}

\item{Z}{An \eqn{tt n x q_k} matrix of independent variables}

\item{niter}{Total number of iterations}

\item{nretain}{Number of iterations to retain (must be smaller than niter)}

\item{W_prior}{List of priors for estimating \eqn{W}, generated by \code{\link{W_priors}}}

\item{rho_prior}{List of priors for estimating \eqn{\rho}, generated by \code{\link{rho_priors}}}

\item{beta_prior}{List of priors for the slope coefficients \eqn{\beta}, generated by \code{\link{beta_priors}}}

\item{sigma_prior}{List of priors for the error variance \eqn{\sigma^2}, generated by \code{\link{sigma_priors}}}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and direct, indirect, and total effects.
}
\description{
The model takes the form \eqn{Y = \rho f(\Omega)Y + X \beta_1 + f(\Omega)X \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,\sigma^2)}
}
\examples{
n = 20; tt = 10
dgp_dat = sim_dgp(n =n, tt = tt, rho = .5, beta1 = c(.5,1),beta2 = c(-1,.5),
            beta3 = c(1.5), sigma2 = .5)
res = sdmw(Y = dgp_dat$Y,tt = tt,X = dgp_dat$X,Z = dgp_dat$Z,niter = 20,nretain = 10)
}
