% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmw.R
\name{sdmw}
\alias{sdmw}
\title{A sampler for estimating the W matrix in a SDM type model}
\usage{
sdmw(
  Y,
  tt,
  X = matrix(0, nrow(Y), 0),
  Z = matrix(1, nrow(Y), 1),
  niter = 1000,
  nretain = 250,
  W_prior = W_priors(n = nrow(Y)/tt),
  rho_prior = rho_priors(),
  beta_prior = beta_priors(k = ncol(X) * 2 + ncol(Z)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{numeric \eqn{tt * n} by \eqn{1} matrix containing the dependent variables.}

\item{tt}{single number greater or equal to 1. Denotes the number of time observations.}

\item{X}{numeric \eqn{tt * n} by \eqn{k_1} design matrix of independent variables. These will be
additionally spatially lagged by \eqn{W}. The default value is a matrix with \eqn{tt * n}
rows and zero columns.}

\item{Z}{numeric \eqn{tt * n} by \eqn{k_3} design matrix of independent variables (which are not
spatially lagged by \eqn{W}). The default value
is a \eqn{tt * n} by \eqn{1} vector of ones (an intercept for the model).}

\item{niter}{single number greater or equal to 1, indicating the total number of draws.
Will be automatically coerced to integer. The default value is 1000.}

\item{nretain}{single number greater or equal to 0, indicating the number of draws
kept after the burn-in. Will be automatically coerced to integer. The default value is 500.}

\item{W_prior}{list containing prior settings for estimating the spatial weight matrix \eqn{W},
generated by the smart constructor \code{\link{W_priors}}.}

\item{rho_prior}{list of prior settings for estimating \eqn{\rho},
generated by the smart constructor \code{\link{rho_priors}}}

\item{beta_prior}{list containing priors for the slope coefficients \eqn{\beta},
generated by the smart constructor \code{\link{beta_priors}}.}

\item{sigma_prior}{list containing priors for the error variance \eqn{\sigma^2},
generated by the smart constructor \code{\link{sigma_priors}}}
}
\value{
List with posterior samples for \eqn{\rho}, \eqn{\beta}, \eqn{\sigma^2}, \eqn{w}, and
average direct, indirect, and total effects.
}
\description{
The model takes the form \eqn{Y = \rho f(\Omega)Y + X \beta_1 + f(\Omega)X \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,I\sigma^2)}
}
\examples{
n = 20; tt = 10
dgp_dat = sim_dgp(n =n, tt = tt, rho = .5, beta1 = c(.5,1),beta2 = c(-1,.5),
            beta3 = c(1.5), sigma2 = .5)
res = sdmw(Y = dgp_dat$Y,tt = tt,X = dgp_dat$X,Z = dgp_dat$Z,niter = 20,nretain = 10)
}
