% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sigma_priors}
\alias{sigma_priors}
\title{Prior Distributions for the error variance}
\usage{
sigma_priors(sigma_rate_prior = 0.1, sigma_shape_prior = 0.1)
}
\arguments{
\item{sigma_rate_prior}{Sigma rate prior parameter (scalar), default: \eqn{.1}}

\item{sigma_shape_prior}{Sigma shape prior parameter (scalar), default: \eqn{.1}

This function allows the user to specify priors for the error variance \eqn{\sigma^2}.}
}
\description{
Prior Distributions for the error variance
}
