% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sigma_priors}
\alias{sigma_priors}
\title{Set prior specification for the error variance using an inverse Gamma distribution}
\usage{
sigma_priors(sigma_rate_prior = 0.001, sigma_shape_prior = 0.001)
}
\arguments{
\item{sigma_rate_prior}{Sigma rate prior parameter (scalar), default: \eqn{0.001}.}

\item{sigma_shape_prior}{Sigma shape prior parameter (scalar), default: \eqn{0.001}.

This function allows the user to specify priors for the error variance \eqn{\sigma^2}.}
}
\description{
Set prior specification for the error variance using an inverse Gamma distribution
}
