% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slxw.R
\name{slxw}
\alias{slxw}
\title{A sampler for estimating the W matrix in an SLX type model}
\usage{
slxw(
  Y,
  tt,
  X = matrix(0, nrow(Y), 0),
  Z = matrix(1, nrow(Y), 1),
  niter = 1000,
  nretain = 250,
  W_prior = W_priors(n = nrow(Y)/tt),
  beta_prior = beta_priors(k = ncol(X) * 2 + ncol(Z)),
  sigma_prior = sigma_priors()
)
}
\arguments{
\item{Y}{numeric \eqn{N \times 1} matrix containing the dependent variables, where \eqn{N} is the number of
spatial (\eqn{n}) times time observations (\eqn{tt=T}). Note that the observations have to be ordered first by
\eqn{n} and then by \eqn{T}.}

\item{tt}{single number greater or equal to 1. Denotes the number of time observations. \eqn{tt = T}.}

\item{X}{numeric \eqn{N \times k_1} design matrix of independent variables. These will be
automatically spatially lagged. If no spatially lagged variable is included in the model
a matrix with \eqn{N} rows and zero columns should be supplied (the default value). Note:
either \eqn{X} or \eqn{Z} has to be a matrix with at least one column.}

\item{Z}{numeric \eqn{N \times k_3} design matrix of independent variables which are not
spatially lagged. The default value is a \eqn{N \times 1} vector of ones (i.e. an intercept for the model).
Note: either \eqn{X} or \eqn{Z} has to be a matrix with at least one column.}

\item{niter}{single number greater or equal to 1, indicating the total number of draws.
Will be automatically coerced to integer. The default value is 1000.}

\item{nretain}{single number greater or equal to 0, indicating the number of draws
kept after the burn-in. Will be automatically coerced to integer. The default value is 250.}

\item{W_prior}{list containing prior settings for estimating the spatial weight matrix \eqn{W},
generated by the smart constructor \code{\link{W_priors}}.}

\item{beta_prior}{list containing priors for the slope coefficients \eqn{\beta},
generated by the smart constructor \code{\link{beta_priors}}. The ordering of the priors is:
(1) priors of \eqn{X}, (2) priors of spatially lagged \eqn{X}, (3) priors of \eqn{Z}.}

\item{sigma_prior}{list containing priors for the error variance \eqn{\sigma^2},
generated by the smart constructor \code{\link{sigma_priors}}}
}
\value{
List with posterior samples for \eqn{\beta_1}, \eqn{\beta_2}, \eqn{\beta_3}, and \eqn{\sigma^2}.
}
\description{
The model takes the form \eqn{Y = X \beta_1 + f(\Omega)X \beta_2 + Z \beta_3 +  \epsilon}, with \eqn{\epsilon \sim N(0,I\sigma^2)}
}
\examples{
set.seed(123)
n = 20; tt = 10
dgp_dat = sim_dgp(n = 20, tt = 10, rho = 0, beta1 = c(1,-1),
                  beta2 = c(0,.5), beta3 = c(.2), sigma2 = .5)
res = slxw(Y = dgp_dat$Y, tt = tt, X = dgp_dat$X, Z = dgp_dat$Z,
                  niter = 20, nretain = 10)
}
